+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/screen/loggedIn/wallet/TopUpMintViewModel.kt
@@ -23,8 +23,8 @@ package com.vitorpamplona.amethyst.ui.screen.loggedIn.wallet
 import androidx.compose.runtime.Immutable
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.vitorpamplona.amethyst.commons.cashu.ops.describeMintError
 import com.vitorpamplona.amethyst.model.nip60Cashu.CashuWalletState
-import com.vitorpamplona.amethyst.model.nip60Cashu.describeMintError
 import com.vitorpamplona.amethyst.ui.screen.loggedIn.AccountViewModel
 import com.vitorpamplona.quartz.nip47WalletConnect.Nip47WalletConnect
 import com.vitorpamplona.quartz.nip47WalletConnect.rpc.PayInvoiceMethod
@@ -47,7 +47,7 @@ import kotlinx.coroutines.launch
  *
  * Reuses the same funding primitives as [ReloadMintViewModel] —
  * [CashuWalletState.rebalance] for a mint-to-mint move and
- * [com.vitorpamplona.amethyst.model.nip60Cashu.CashuWalletOps.startMintFromLightning] /
+ * [com.vitorpamplona.amethyst.commons.cashu.ops.CashuWalletOps.startMintFromLightning] /
  * `completeMintFromLightning` for an LN top-up — but without any of the
  * zap-specific machinery (recipient, shared-mint intersection, fixed send
  * amount + shortfall, terminal nutzap, fund-then-send atomicity). The user
